requirements.txt
requests==2.9.1
pyserial==2.6
